@@ -3,8 +3,7 @@ description: A flutter package that has an animated search bar with loads of cus
 version: 2.1.3
 homepage: https://github.com/Imgkl/anim_search_bar
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: "^3.5.1"
 
 dependencies:
   flutter:
