@@ -1,4 +1,5 @@
 import 'dart:math';
+import 'package:anim_search_bar/src/consts.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
@@ -26,9 +27,10 @@ class AnimSearchBar extends StatefulWidget {
   final TextEditingController textController;
   final Icon? suffixIcon;
   final Icon? prefixIcon;
-  final String helpText;
+  final Icon backIcon;
+  final String? helpText;
   final int animationDurationInMilli;
-  final onSuffixTap;
+  final VoidCallback? onSuffixTap;
   final bool rtl;
   final bool autoFocus;
   final TextStyle? style;
@@ -38,108 +40,116 @@ class AnimSearchBar extends StatefulWidget {
   final Color? searchIconColor;
   final Color? textFieldIconColor;
   final List<TextInputFormatter>? inputFormatters;
-  final bool boxShadow;
+  final List<BoxShadow>? boxShadow;
   final TextInputAction textInputAction;
-  final Function(int) searchBarOpen;
+  final void Function(bool)? searchBarOpen;
   final void Function(String newVal)? onChanged;
   final bool clearOnSuffixTap;
   final bool clearOnClose;
   final bool closeOnSubmit;
 
   const AnimSearchBar({
-    Key? key,
+    super.key,
 
-    /// The width cannot be null
+    // The width cannot be null
     required this.width,
-    required this.searchBarOpen,
-    /// The textController cannot be null
+    this.searchBarOpen,
+    // The textController cannot be null
     required this.textController,
-    this.suffixIcon,
+    this.suffixIcon = suffixIconBlack,
     this.prefixIcon,
-    this.helpText = "Search...",
-    
-    /// Height of wrapper container
+    this.helpText,
+    this.backIcon = const Icon(Icons.arrow_back_ios_new_outlined),
+    // Height of wrapper container
     this.height = 100,
     
-    /// choose your custom color
+    // choose your custom color
     this.color = Colors.white,
 
-    /// choose your custom color for the search when it is expanded
+    // choose your custom color for the search when it is expanded
     this.textFieldColor = Colors.white,
 
-    /// choose your custom color for the search when it is expanded
+    // choose your custom color for the search when it is expanded
     this.searchIconColor = Colors.black,
 
-    /// choose your custom color for the search when it is expanded
+    // choose your custom color for the search when it is expanded
     this.textFieldIconColor = Colors.black,
     this.textInputAction = TextInputAction.done,
 
-    /// The onSuffixTap cannot be null
-    required this.onSuffixTap,
+    // The onSuffixTap cannot be null
+    this.onSuffixTap,
     this.animationDurationInMilli = 375,
 
-    
-    /// make the search bar to open from right to left
+  
+    // make the search bar to open from right to left
     this.rtl = false,
 
-    /// make the keyboard to show automatically when the searchbar is expanded
-    this.autoFocus = false,
+    // make the keyboard to show automatically when the searchbar is expanded
+    this.autoFocus = true,
 
-    /// TextStyle of the contents inside the searchbar
+    // TextStyle of the contents inside the searchbar
     this.style,
 
-    /// close the search on suffix tap
+    // close the search on suffix tap
     this.closeSearchOnSuffixTap = false,
 
-    /// enable/disable the box shadow decoration
-    this.boxShadow = true,
+    // Textfield boxShadow
+    this.boxShadow = boxShadowBlack,
 
-    /// can add list of inputformatters to control the input
+    // can add list of inputformatters to control the input
     this.inputFormatters,
 
     this.onChanged,
     this.clearOnSuffixTap = false,
-    this.clearOnClose = false,
+    this.clearOnClose = true,
     this.closeOnSubmit = true,
-  }) : super(key: key);
+  });
 
   @override
-  _AnimSearchBarState createState() => _AnimSearchBarState();
+  AnimSearchBarState createState() => AnimSearchBarState();
 }
 
 
 
-class _AnimSearchBarState extends State<AnimSearchBar>
+class AnimSearchBarState extends State<AnimSearchBar>
     with SingleTickerProviderStateMixin {
-      ///toggle - 0 => false or closed
-      ///toggle 1 => true or open
-        int toggle = 0;
 
-        /// * use this variable to check current text from OnChange
-        String textFieldValue = '';
+  bool isSearchbarOpen = false;
 
-
-  ///initializing the AnimationController
-  late AnimationController _con;
-  FocusNode focusNode = FocusNode();
+  //initializing the AnimationController
+  late AnimationController _animationController;
+  final FocusNode focusNode = FocusNode();
 
   @override
   void initState() {
     super.initState();
 
-    ///Initializing the animationController which is responsible for the expanding and shrinking of the search bar
-    _con = AnimationController(
+    //Initializing the animationController which is responsible for the expanding and shrinking of the search bar
+    _animationController = AnimationController(
       vsync: this,
 
-      /// animationDurationInMilli is optional, the default value is 375
+      // animationDurationInMilli is optional, the default value is 375
       duration: Duration(milliseconds: widget.animationDurationInMilli),
     );
   }
 
-  unfocusKeyboard() {
+  void unfocusKeyboard() {
     final FocusScopeNode currentScope = FocusScope.of(context);
     if (!currentScope.hasPrimaryFocus && currentScope.hasFocus) {
       FocusManager.instance.primaryFocus?.unfocus();
+    }
+  }
+
+  void toogleSearchbar(){
+    setState(() {
+      isSearchbarOpen = !isSearchbarOpen;
+    });
+  }
+
+  void closeOnSubmit(){
+    if(widget.closeOnSubmit) {
+      unfocusKeyboard();
+      toogleSearchbar();                    
     }
   }
 
@@ -148,166 +158,116 @@ class _AnimSearchBarState extends State<AnimSearchBar>
     return Container(
       height: widget.height,
 
-      ///if the rtl is true, search bar will be from right to left
-      alignment: widget.rtl ? Alignment.centerRight : Alignment(-1.0, 0.0),
+      //if the rtl is true, search bar will be from right to left
+      alignment: widget.rtl ? Alignment.centerRight : Alignment.centerLeft,
 
-      ///Using Animated container to expand and shrink the widget
+      //Using Animated container to expand and shrink the widget
       child: AnimatedContainer(
         duration: Duration(milliseconds: widget.animationDurationInMilli),
         height: widget.height,
-        width: (toggle == 0) ? 48.0 : widget.width,
+        width: isSearchbarOpen ? widget.width : 48,
         curve: Curves.easeOut,
         decoration: BoxDecoration(
-          /// can add custom  color or the color will be white
-          color: toggle == 1 ? widget.textFieldColor : widget.color,
+          // can add custom  color or the color will be white
+          color: isSearchbarOpen ? widget.textFieldColor : widget.color,
           borderRadius: BorderRadius.circular(30.0),
 
-          /// show boxShadow unless false was passed
-          boxShadow: !widget.boxShadow
-              ? null
-              : [
-                  BoxShadow(
-                    color: Colors.black26,
-                    spreadRadius: -10.0,
-                    blurRadius: 10.0,
-                    offset: Offset(0.0, 10.0),
-                  ),
-                ],
+          // show boxShadow unless false was passed
+          boxShadow: widget.boxShadow
         ),
         child: Stack(
           children: [
-            ///Using Animated Positioned widget to expand and shrink the widget
+            //Using Animated Positioned widget to expand and shrink the widget
             AnimatedPositioned(
               duration: Duration(milliseconds: widget.animationDurationInMilli),
               top: 6.0,
               right: 7.0,
               curve: Curves.easeOut,
               child: AnimatedOpacity(
-                opacity: (toggle == 0) ? 0.0 : 1.0,
-                duration: Duration(milliseconds: 200),
+                opacity: isSearchbarOpen ? 1.0 : 0.0,
+                duration: const Duration(milliseconds: 200),
                 child: Container(
-                  padding: EdgeInsets.all(8.0),
+                  padding: const EdgeInsets.all(8.0),
                   decoration: BoxDecoration(
-                    /// can add custom color or the color will be white
                     color: widget.color,
                     borderRadius: BorderRadius.circular(30.0),
                   ),
                   child: AnimatedBuilder(
-                    child: GestureDetector(
-                      onTap: () {
-                        try {
-                          ///trying to execute the onSuffixTap function
-                          widget.onSuffixTap();
-
-                          // * if field empty then the user trying to close bar
-                          if (textFieldValue == '') {
-                            unfocusKeyboard();
-                            setState(() {
-                              toggle = 0;
-                            });
-                            widget.searchBarOpen(toggle);
-
-                            ///reverse == close
-                            _con.reverse();
-                          }
-
-                          // // * why not clear textfield here?
-                          if (widget.clearOnSuffixTap) {
-                          widget.textController.clear();
-                          textFieldValue = '';
-                          widget.onChanged?.call("");
-                        }
-
-                          ///closeSearchOnSuffixTap will execute if it's true
-                          if (widget.closeSearchOnSuffixTap) {
-                            unfocusKeyboard();
-                            setState(() {
-                              toggle = 0;
-                            });
-                          }
-                        } catch (e) {
-                          ///print the error if the try block fails
-                          print(e);
-                        }
-                      },
-
-                      ///suffixIcon is of type Icon
-                      child: widget.suffixIcon != null
-                          ? widget.suffixIcon
-                          : Icon(
-                              Icons.close,
-                              size: 20.0,
-                              color: widget.textFieldIconColor,
-                            ),
-                    ),
                     builder: (context, widget) {
-                      ///Using Transform.rotate to rotate the suffix icon when it gets expanded
+                      //Using Transform.rotate to rotate the suffix icon when it gets expanded
                       return Transform.rotate(
-                        angle: _con.value * 2.0 * pi,
+                        angle: _animationController.value * 2.0 * pi,
                         child: widget,
                       );
                     },
-                    animation: _con,
+                    animation: _animationController,
+                    child: GestureDetector(
+                      onTap: () {
+                        //trying to execute the onSuffixTap function
+                        if(widget.onSuffixTap != null) widget.onSuffixTap!();
+                        // * if field empty then the user trying to close bar
+                        if (widget.textController.text == '') {
+                          unfocusKeyboard();
+                          toogleSearchbar();
+                          if(widget.searchBarOpen != null)widget.searchBarOpen!(isSearchbarOpen);
+                          ///reverse == close
+                          _animationController.reverse();
+                        }
+                        // // * why not clear textfield here?
+                        if (widget.clearOnSuffixTap) {
+                        widget.textController.clear();
+                        widget.onChanged?.call("");
+                      }
+                        ///closeSearchOnSuffixTap will execute if it's true
+                        if (widget.closeSearchOnSuffixTap) {
+                          unfocusKeyboard();
+                          toogleSearchbar();
+                        }
+                      },
+                      child: widget.suffixIcon,
+                    ),
                   ),
                 ),
               ),
             ),
             AnimatedPositioned(
               duration: Duration(milliseconds: widget.animationDurationInMilli),
-              left: (toggle == 0) ? 20.0 : 40.0,
+              left: isSearchbarOpen ? 40.0 : 20.0,
               curve: Curves.easeOut,
               top: 11.0,
 
-              ///Using Animated opacity to change the opacity of th textField while expanding
+              //Using Animated opacity to change the opacity of th textField while expanding
               child: AnimatedOpacity(
-                opacity: (toggle == 0) ? 0.0 : 1.0,
-                duration: Duration(milliseconds: 200),
+                opacity: isSearchbarOpen ? 1.0 : 0.0,
+                duration: const Duration(milliseconds: 200),
                 child: Container(
                   padding: const EdgeInsets.only(left: 10),
                   alignment: Alignment.topCenter,
                   width: widget.width / 1.7,
                   child: TextField(
-                    ///Text Controller. you can manipulate the text inside this textField by calling this controller.
+                    //Text Controller. you can manipulate the text inside this textField by calling this controller.
                     controller: widget.textController,
                     inputFormatters: widget.inputFormatters,
                     focusNode: focusNode,
                     textInputAction: widget.textInputAction,
-                    cursorRadius: Radius.circular(10.0),
+                    cursorRadius: const Radius.circular(10.0),
                     cursorWidth: 2.0,
                     onChanged: (value) {
-                      textFieldValue = value;
                       widget.onChanged?.call(value);
                     },
                     onSubmitted: (value) {
-                     if(widget.closeOnSubmit) {
-                        unfocusKeyboard();
-                        setState(() {
-                          toggle = 0;
-                        });
-                     }
-                      // widget.textController.clear(),
+                      closeOnSubmit();
                     },
-                    onEditingComplete: () {
-                      /// on editing complete the keyboard will be closed and the search bar will be closed
-                     if(widget.closeOnSubmit) {
-                        unfocusKeyboard();
-                        setState(() {
-                          toggle = 0;
-                        });
-                     }
-                    },
+                    onEditingComplete: closeOnSubmit,
 
-                    ///style is of type TextStyle, the default is just a color black
-                    style: widget.style != null
-                        ? widget.style
-                        : TextStyle(color: Colors.black),
+                    style: widget.style ?? const TextStyle(color: Colors.black),
                     cursorColor: Colors.black,
                     decoration: InputDecoration(
                       contentPadding: const EdgeInsets.only(bottom: 5),
                       isDense: true,
                       floatingLabelBehavior: FloatingLabelBehavior.never,
                       labelText: widget.helpText,
-                      labelStyle: TextStyle(
+                      labelStyle: const TextStyle(
                         color: Color(0xff5B5B5B),
                         fontSize: 17.0,
                         fontWeight: FontWeight.w500,
@@ -327,63 +287,48 @@ class _AnimSearchBarState extends State<AnimSearchBar>
             Material(
               /// can add custom color or the color will be white
               /// toggle button color based on toggle state
-              color: toggle == 0 ? widget.color : widget.textFieldColor,
+              color: isSearchbarOpen ?  widget.textFieldColor : widget.color ,
               borderRadius: BorderRadius.circular(30.0),
               child: IconButton(
                 splashRadius: 19.0,
 
-                ///if toggle is 1, which means it's open. so show the back icon, which will close it.
-                ///if the toggle is 0, which means it's closed, so tapping on it will expand the widget.
                 ///prefixIcon is of type Icon
                 icon: widget.prefixIcon != null
-                    ? toggle == 1
-                        ? Icon(
-                            Icons.arrow_back_ios,
-                            color: widget.textFieldIconColor,
-                          )
-                        : widget.prefixIcon!
+                    ? isSearchbarOpen
+                      ? widget.backIcon
+                      : widget.prefixIcon!
                     : Icon(
-                        toggle == 1 ? Icons.arrow_back_ios : Icons.search,
+                        isSearchbarOpen ? Icons.arrow_back_ios : Icons.search,
                         // search icon color when closed
-                        color: toggle == 0
-                            ? widget.searchIconColor
-                            : widget.textFieldIconColor,
+                        color: isSearchbarOpen
+                          ? widget.textFieldIconColor
+                          : widget.searchIconColor,
                         size: 20.0,
                       ),
                 onPressed: () {
                   setState(
                     () {
                       ///if the search bar is closed
-                      if (toggle == 0) {
-                        toggle = 1;
-                        setState(() {
-                          ///if the autoFocus is true, the keyboard will pop open, automatically
-                          if (widget.autoFocus)
-                            FocusScope.of(context).requestFocus(focusNode);
-                        });
+                      if (!isSearchbarOpen) {
+                        isSearchbarOpen = true;
+                        if (widget.autoFocus) FocusScope.of(context).requestFocus(focusNode);
 
                         ///forward == expand
-                        _con.forward();
+                        _animationController.forward();
                       } else {
                         ///if the search bar is expanded
-                        toggle = 0;
+                        isSearchbarOpen = false;
 
                         ///if the autoFocus is true, the keyboard will close, automatically
-                        setState(() {
-                          if (widget.autoFocus) unfocusKeyboard();
-                        });
+                        if (widget.autoFocus) unfocusKeyboard();
 
                         ///reverse == close
-                        _con.reverse();
+                        _animationController.reverse();
                       }
                     },
-
                   );
-                  if (toggle == 0 &&  widget.clearOnClose) {
-                          widget.textController.clear();
-                          textFieldValue = '';
-                        }
-                  widget.searchBarOpen(toggle);
+                  if (!isSearchbarOpen  &&  widget.clearOnClose) widget.textController.clear();
+                  if(widget.searchBarOpen != null) widget.searchBarOpen!(isSearchbarOpen);
                 },
               ),
             ),
@@ -393,3 +338,4 @@ class _AnimSearchBarState extends State<AnimSearchBar>
     );
   }
 }
+
